requirements.txt
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
bidict==0.21.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
colorama==0.4.4
fastapi==0.74.1
Flask==1.1.4
Flask-SocketIO==5.1.1
h11==0.13.0
idna==3.3
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.28
passlib==1.7.4
protobuf==3.19.4
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
python-engineio==4.3.1
python-socketio==5.5.2
requests==2.27.1
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
starlette==0.17.1
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.5
Werkzeug==1.0.1
